main_bot/requirements.txt
aiocache==0.12.3
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiolimiter==1.2.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asyncio==4.0.0
attrs==25.3.0
bitarray==3.7.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
dnspython==2.7.0
fastapi==0.116.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
motor==3.7.1
multidict==6.6.4
propcache==0.3.2
pyaes==1.6.1
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.14.1
PyNaCl==1.5.0
pyrogrammod==2.2.1
PySocks==1.7.1
pytoniq-core==0.1.44
requests==2.32.5
setuptools==69.1.1
sniffio==1.3.1
starlette==0.47.3
TgCrypto==1.2.5
tonsdk==1.0.15
tonutils==0.5.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
wheel==0.43.0
x25519==0.0.2
yarl==1.20.1
